Moved task utils + removed references to input in function
