+++ lightworks/sdk/utils/__init__.py
@@ -32,3 +32,4 @@ from .random_utils import (
     random_permutation,
     random_unitary,
 )
+from .task_utils import validate_states
